+++ backend/models/PurchaseReturn.js
@@ -6,9 +6,9 @@ const PurchaseReturn = sequelize.define("PurchaseReturn", {
   originalPoNo: { type: DataTypes.STRING },
   supplierName: { type: DataTypes.STRING, allowNull: false },
   items:        { type: DataTypes.JSON },
-  subtotal:     { type: DataTypes.FLOAT, defaultValue: 0 },
-  gstAmount:    { type: DataTypes.FLOAT, defaultValue: 0 },
-  totalAmount:  { type: DataTypes.FLOAT, defaultValue: 0 },
+  subtotal:     { type: DataTypes.DECIMAL(15, 2), defaultValue: 0 },
+  gstAmount:    { type: DataTypes.DECIMAL(15, 2), defaultValue: 0 },
+  totalAmount:  { type: DataTypes.DECIMAL(15, 2), defaultValue: 0 },
   reason:       { type: DataTypes.STRING },
   status:       { type: DataTypes.ENUM("pending","approved"), defaultValue: "approved" },
 });
