+++ backend/models/SellThrough.js
@@ -6,7 +6,7 @@ const SellThrough = sequelize.define("SellThrough", {
   name:               { type: DataTypes.STRING, allowNull: false },
   stockAtDistributor: { type: DataTypes.INTEGER, defaultValue: 0 },
   sold30d:            { type: DataTypes.INTEGER, defaultValue: 0 },
-  velocity:           { type: DataTypes.FLOAT, defaultValue: 0 },
+  velocity:           { type: DataTypes.DECIMAL(15, 2), defaultValue: 0 },
   trend:              { type: DataTypes.STRING, defaultValue: "stable" }
 });
