+++ backend/models/SupplierScheme.js
@@ -9,7 +9,7 @@ const SupplierScheme = sequelize.define("SupplierScheme", {
   applicableItems:{ type: DataTypes.JSON, defaultValue: [] },
   minQty:         { type: DataTypes.INTEGER, defaultValue: 0 },
   freeQty:        { type: DataTypes.INTEGER, defaultValue: 0 },
-  discountPercent:{ type: DataTypes.FLOAT, defaultValue: 0 },
+  discountPercent:{ type: DataTypes.DECIMAL(15, 2), defaultValue: 0 },
   slabRates:      { type: DataTypes.JSON, defaultValue: [] },
   validFrom:      { type: DataTypes.DATEONLY },
   validTo:        { type: DataTypes.DATEONLY },
